Pin an individual action to its corresponding SHA

Pin an individual action to its corresponding SHA

Usage

pin_action(action = "r-lib/actions/check-r-package@v2")

Arguments

action

the action written with github syntax that contains at least the github user, repo, and tag

Value

the same action with an SHA and a comment that indicates the tag the sha belongs to (this may be different than the tag if you use a floating tag)

Examples

pin_action("r-lib/actions/check-r-package@v2")
#> [1] "r-lib/actions/check-r-package@bd49c52ffe281809afa6f0fecbf37483c5dd0b93 #v2.11.3"